requirements.txt
langchain~=0.0.321
pinecone-client==2.1.0
streamlit_chat
openai==0.27.8
tiktoken
